requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
